:coffin: remove dead code
