+++ src/analyzer.rs
@@ -10,6 +10,7 @@ use crate::analyzer::kde::BenchmarkKde;
 use crate::benchmark_wrapper::BenchmarkType;
 use crate::analyzer::linear_model::LinearModel;
 use crate::benchmark_wrapper::PerformanceBenchmark;
+use crate::use_model::CsvLine;
 
 use serde::{Deserialize, Serialize};
 use serde_json::json;
@@ -60,5 +61,43 @@ impl Analysis {
         write!(output, "{}", Self::all_to_json(xs))?;
 
         Ok(())
+    }
+
+    pub fn find_lowest_upper_bound(xs: Vec<Self>, line: &CsvLine) -> Option<&LinearModel> {
+        // TODO: from LinearModels, recode me
+        /*
+    pub fn find_lowest_upper_bound(&self, line: &CsvLine) -> Option<&LinearModelJSON> {
+        let mut res = None;
+        for lm in self.0.iter() {
+            // Apples and oranges
+            if lm.is_read_op != (line.io_type == 'r') {
+                continue;
+            }
+            println!("{:?}", lm);
+
+            let approximated_time = lm.model.evaluate(line.bytes);
+            println!("{:?} -> {}", lm, approximated_time);
+
+            // we are looking for an upper bound. Thus if it is lower, we can instantly reject it.
+            if approximated_time < line.sec {
+                continue;
+            }
+
+            // do we have a upper bound already?
+            res = match res {
+                // if not, this is the best until now
+                None => Some(lm),
+                // if so, lets choose the tighter bound
+                Some(lm2) => Some(if lm2.model.evaluate(line.bytes) < approximated_time {
+                    lm2
+                } else {
+                    lm
+                }),
+            }
+        }
+        res
+    }
+        */
+        None
     }
 }
