+++ src/subprograms/use_model.rs
@@ -1,7 +1,6 @@
 use serde::Deserialize;
 
 #[derive(Debug, Deserialize)]
-#[allow(dead_code)]
 pub struct CsvLine {
     pub filename: String,
     pub io_type: char,
